Return structured object with score and points
Points has slightly non trivial logic. Since we're computing it in code already, we might as well return it to the backend as well.

So this refactors the `updateScore` and `updatePoints` logic to just compute those totals as well as value per hole and return everything.

Then the submit function can reuse that logic and return everything to the backend in a nice object that has named fields instead of just lists of strings.