Adds "Team Points" display.
Instead of recomputing each players point, we can just read the `points` div and add them up.

We trigger when we recompute a single player's points, and since thats the only one that has changed, we update that first, then recompute the overall total from reading all of the fields.