Configurable insert rows into the table.
Unfortunately my lack of experience in JavaScript is showing here.

I tried to do a "functional" style generation of the table. But I haven't found a way to extract a particular row out of the table by ID; the only thing I see is document.getElementById, but for that to work the table has to be added to the document first.

So now we're just mixing functional stuff that generates the table, and then later functions that modify that table.

Coming next, score and point stuff for these partial results.