General score,points,net and TOTAL as "Results"
We have the "Result" concept and it should apply uniformly to any subset of the holes, front 9, back 9, full 18, etc.

This change creates rows for configured result.

It does this by injecting them directly into the score table. Its not a tested table in order for the columns to all line up.

There were some hacks to get the ordering to be correct. We're always inserting after the Nth hole, so we insert things in reverse to make sure the final output is the way we want.

This change does NOT wire these new elements up. That comes later. As such, I've left the existing wired up fields around.