Simplify output to only players and score.
All of the extra fields like points and net and totals are nice UI features *because* they're reactive as hole scores are entered or changed.

Once we send it to the spreadsheet they're static, so once someone comes along and corrects a number, all of that computation is invalid.

Since people are always going to want to manually tweak this stuff, we'll write only the essentials.

- date
- courseName
- teamName
- playerName
- hole1Score
- hole2Score
- ...
- holeNScore

(one row per player)