Default to sending the data to AppScript
So far the workflow is to copy/paste from git into AppScript.

This saves us the step of having to remember to uncomment out the line to actually write to the spreadsheet.

When running locally, you'll see

```
Uncaught ReferenceError: google is not defined
    at submit (index.html:426:9)
    at HTMLButtonElement.onclick (index.html:24:32)
submit @ index.html:426
onclick @ index.html:24
```

But its the web! So it doesn't actually matter.

Replaced the alert with a log, so debugging is simple as well