Store scores in localStorage; adds clear button
We expect users to enter this on their phone, probably opening the link from email into that "ephemeral" browser, so if they navigate away, the page will reload. There are a lot of fields to enter, so that's pretty annoying.

So we use localStorage to store every hole score the user enters. They are repopulated on page refresh.

The `clear` button clears the fields and the cache.

Known issues:
- Populating saved scores doesn't update the Results.
- Clearing entered scores doesn't clear the Results.

I think we can address both with the same fix, a method to recompute all of it.

Also known:
- Doesn't store team name, player names, nor target.

This is because these are no longer human controlled fields, those will be done by team drop down and config provided.