Restrict hole scores to between min and max.
By default it seems that the min and max rules on an input field don't actually apply to typed values.

So we enforce it ourselves. Despite being a number field, the `value` field returns a string. So just using comparisions does *string* comparision. That is "66" < "7". Math.max and Math.min seem to conver to numbers first.