Shrink width of input boxes and remove empty space.
The width of the column is constrained by the width of the largest item, which is the names.

I define width in terms of characters as opposed to pixels to try to make sure it looks the same regardless of resolution. Im not sure if pixels are abstracted away, does a fancy phone have a smaller pixel than a cheap phone?