Submit button no longer clears, only locks input.
The idea is that users like to be able to see the scores they've entered. They might want to screenshot it or copy it or something.

I've updated the popup message as well.

The clear button remains enabled.

Refreshing the page, unlocks all of the input fields. Since the scores haven't been cleared, they are presented with a completely full scorecard.