Export computed values to spreadsheet.
Adds target and results to each players output, and a new row for team level values. Results are added in the order defined in the config with values appearing as they do in the UI: score, points, net.

We found that trying to recompute these values in the spreadsheet was too complicated, so we'll just keep what the frontend has already computed. (Obviously not resistant to tampering). The main downside here is you can't *correct* scores or targets in the spreadsheet; or at least changing any value will **NOT** update anything else.