Use `oninput` for `Target` and `Score` inputs.
This changes values immediately, instead of when the input box loses focus. So if a user is typing, they can immediately see the results of their new value.

Because the typical flow is likely to only click on the next hole after you've completed it, you're always an entire hole behind on the stats.

This does NOT change the function for the teams field, but that is just a drop-down so it doesn't have the same problem.