Functions to save and load scores into sheets.
The local storage within the browser isn't working for some users. We haven't been able to figure out why. Reloading the page or navigating away and loosing all of your progress is very frustrating. So a solution is to save the scores in the backend. To keep things simple, we'll have this explicitly triggered by the user via `Save` and `Load` buttons.

This change creates the two functions on the backend to do the saving and loading.

I created a new spreadsheet to be this storage. Each team will get it's own sheet, and every write operation will clear the entire sheet. Every read operation will just grab everything. We'll rely on the frontend to validate and populate appropriately. This approach means we don't have to do any complicated filtering or modification of rows.